#2003: extract HA liveness slots into bpf_map/ha.rs
Pure code motion. afxdp/bpf_map/mod.rs (~1063 LOC) mixes several
concerns; split it into focused submodules following the existing
publish_conntrack.rs precedent (#1356). This first step lifts the
redundancy-group liveness-slot cluster — the XSK-slot and per-binding
heartbeat-slot map writes the XDP shim consults to decide whether a
binding is active — into a new bpf_map/ha.rs.

Functions moved verbatim (no signature or logic change):
register_xsk_slot, update_heartbeat_slot, delete_xsk_slot,
delete_heartbeat_slot, maybe_touch_heartbeat, touch_heartbeat,
heartbeat_fresh. Their original pub(super) visibility (= pub(in
crate::afxdp)) is preserved exactly via an explicit pub(in crate::afxdp)
declaration, and mod.rs re-exports them with `pub(in crate::afxdp) use
ha::*;` so the parent afxdp glob (`use self::bpf_map::*`) and every bare-
name call site resolve unchanged.

Validation: cargo build --release and cargo test --release both clean
with this intermediate state (only ha.rs split so far); the metrics and
pin clusters follow in the next two commits.
