+++ userspace-dp/src/afxdp/bpf_map/mod.rs
@@ -86,137 +86,6 @@ pub(super) fn diagnose_raw_ring_state(
     ))
 }
 
-pub(super) fn register_xsk_slot(map_fd: c_int, slot: u32, sock_fd: c_int) -> io::Result<()> {
-    let rc = unsafe {
-        libbpf_sys::bpf_map_update_elem(
-            map_fd,
-            (&slot as *const u32).cast::<c_void>(),
-            (&sock_fd as *const c_int).cast::<c_void>(),
-            libbpf_sys::BPF_ANY as u64,
-        )
-    };
-    if rc != 0 {
-        return Err(io::Error::last_os_error());
-    }
-    Ok(())
-}
-
-pub(super) fn update_heartbeat_slot(map_fd: c_int, slot: u32, timestamp_ns: u64) -> io::Result<()> {
-    let rc = unsafe {
-        libbpf_sys::bpf_map_update_elem(
-            map_fd,
-            (&slot as *const u32).cast::<c_void>(),
-            (&timestamp_ns as *const u64).cast::<c_void>(),
-            libbpf_sys::BPF_ANY as u64,
-        )
-    };
-    if rc != 0 {
-        return Err(io::Error::last_os_error());
-    }
-    Ok(())
-}
-
-pub(super) fn delete_xsk_slot(map_fd: c_int, slot: u32) -> io::Result<()> {
-    let rc =
-        unsafe { libbpf_sys::bpf_map_delete_elem(map_fd, (&slot as *const u32).cast::<c_void>()) };
-    if rc != 0 {
-        return Err(io::Error::last_os_error());
-    }
-    Ok(())
-}
-
-pub(super) fn delete_heartbeat_slot(map_fd: c_int, slot: u32) -> io::Result<()> {
-    let rc =
-        unsafe { libbpf_sys::bpf_map_delete_elem(map_fd, (&slot as *const u32).cast::<c_void>()) };
-    if rc != 0 {
-        return Err(io::Error::last_os_error());
-    }
-    Ok(())
-}
-
-pub(super) fn maybe_touch_heartbeat(binding: &mut BindingWorker, now_ns: u64) {
-    // The fill ring is primed before bind, so the driver's initial NAPI
-    // posts WQEs immediately. All queues should be ready to receive.
-    // No xsk_rx_confirmed gating needed.
-    let age_ns = now_ns.saturating_sub(binding.timers.last_heartbeat_update_ns);
-    if age_ns < HEARTBEAT_UPDATE_INTERVAL_NS {
-        return;
-    }
-    match touch_heartbeat(
-        binding.bpf_maps.heartbeat_map_fd,
-        binding.slot,
-        &binding.live,
-        now_ns,
-    ) {
-        Ok(()) => {
-            if cfg!(feature = "debug-log") {
-                thread_local! {
-                    static HB_LOG_COUNT: std::cell::Cell<u32> = const { std::cell::Cell::new(0) };
-                }
-                let age_ms = age_ns / 1_000_000;
-                if age_ms > 1000 {
-                    debug_log!(
-                        "HB_UPDATE slot={} fd={} age={}ms now_ns={} LATE",
-                        binding.slot,
-                        binding.bpf_maps.heartbeat_map_fd,
-                        age_ms,
-                        now_ns,
-                    );
-                }
-                HB_LOG_COUNT.with(|c| {
-                    let n = c.get();
-                    if n < 5 {
-                        c.set(n + 1);
-                        debug_log!(
-                            "HB_UPDATE[{}] slot={} fd={} age={}ms now_ns={} OK",
-                            n,
-                            binding.slot,
-                            binding.bpf_maps.heartbeat_map_fd,
-                            age_ms,
-                            now_ns,
-                        );
-                    }
-                });
-            }
-            binding.timers.last_heartbeat_update_ns = now_ns;
-        }
-        Err(err) => {
-            eprintln!(
-                "HB_UPDATE_ERR slot={} fd={} age={}ms err={}",
-                binding.slot,
-                binding.bpf_maps.heartbeat_map_fd,
-                age_ns / 1_000_000,
-                err,
-            );
-            binding
-                .live
-                .set_error(format!("update heartbeat slot: {err}"));
-        }
-    }
-}
-
-pub(super) fn touch_heartbeat(
-    map_fd: c_int,
-    slot: u32,
-    live: &BindingLiveState,
-    now_ns: u64,
-) -> io::Result<()> {
-    update_heartbeat_slot(map_fd, slot, now_ns)?;
-    live.set_last_heartbeat_at(now_ns);
-    Ok(())
-}
-
-pub(super) fn heartbeat_fresh(last_heartbeat: Option<chrono::DateTime<Utc>>) -> bool {
-    match last_heartbeat {
-        Some(last) => Utc::now()
-            .signed_duration_since(last)
-            .to_std()
-            .map(|age| age <= HEARTBEAT_STALE_AFTER)
-            .unwrap_or(true),
-        None => false,
-    }
-}
-
 pub(super) struct OwnedFd {
     pub(super) fd: c_int,
 }
@@ -1073,3 +942,13 @@ pub(super) fn delete_session_map_entry_for_removed_session_with_origin(
 mod tests;
 
 mod publish_conntrack;
+
+// #2003: behaviour-preserving code motion. The HA liveness-slot cluster
+// (XSK + heartbeat map writes that gate active-binding state) lives in
+// `ha.rs`; items keep their original `pub(in crate::afxdp)` visibility and
+// are re-exported so the parent `afxdp` glob (`use self::bpf_map::*`)
+// resolves them by bare name exactly as before. Mirrors the
+// `publish_conntrack.rs` precedent (#1356).
+mod ha;
+
+pub(in crate::afxdp) use ha::*;
